fixed dependencies and moved code into the webpage folder
