+++ package.json
@@ -16,13 +16,14 @@
     "@apollo/client": "^3.11.8",
     "@nextui-org/react": "^2.4.8",
     "@octokit/rest": "^21.0.2",
+    "@types/react": "18.3.10",
     "axios": "^1.7.7",
     "dotenv": "^16.4.5",
     "es6-promise": "^4.2.8",
     "isomorphic-fetch": "^3.0.0",
     "isomorphic-git": "^1.27.1",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1",
+    "next": "14.2.13",
     "ts-node": "^10.9.2",
     "url": "^0.11.4",
     "uuid": "^10.0.0"
